Removes the RECOMMENDED section from 1.10
The reason is that it is not meaningful, as the same Subscriber
could be subscribed to several *different* Publishers at the same time.
The spec already requires that a Subscriber is only subscribed to one Publisher,
and only once. And the spec also is clear that the only legal way to reject a
Subscription in a Publisher is by issuing an `onError` signal to it.